requirements.txt
# ─────────────────────────────────────────────
# CORE FRAMEWORK
# ─────────────────────────────────────────────
fastapi
uvicorn[standard]
streamlit

# ─────────────────────────────────────────────
# DATA
# ─────────────────────────────────────────────
pandas
pydantic

# ─────────────────────────────────────────────
# HTTP
# ─────────────────────────────────────────────
requests
httpx
aiohttp

# ─────────────────────────────────────────────
# LLM
# ─────────────────────────────────────────────
groq

# ─────────────────────────────────────────────
# MEMORY
# ─────────────────────────────────────────────
hindsight-client

# ─────────────────────────────────────────────
# ENVIRONMENT
# ─────────────────────────────────────────────
python-dotenv

# ─────────────────────────────────────────────
# TESTING
# ─────────────────────────────────────────────
pytest